refactor: 모든 컨트롤러가 Controller를 구현하는게 아닌 AbstractController를 상속하도록 수정
