+++ tomcat/src/main/java/nextstep/jwp/controller/page/HelloWorldController.java
@@ -1,5 +1,6 @@
 package nextstep.jwp.controller.page;
 
+import nextstep.jwp.controller.AbstractController;
 import nextstep.jwp.controller.Controller;
 import org.apache.coyote.http11.common.HttpHeaders;
 import org.apache.coyote.http11.common.HttpStatus;
@@ -7,19 +8,19 @@ import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.ResponseStatusLine;
 
-public class HelloWorldPageController implements Controller {
+public class HelloWorldController extends AbstractController {
 
     private static final String HELLO_WORLD = "Hello world!";
 
-    private HelloWorldPageController() {
+    private HelloWorldController() {
     }
 
     public static Controller create() {
-        return new HelloWorldPageController();
+        return new HelloWorldController();
     }
 
     @Override
-    public HttpResponse process(final HttpRequest request) {
+    protected HttpResponse doGet(final HttpRequest request) {
         return new HttpResponse(ResponseStatusLine.create(HttpStatus.OK),
                 HttpHeaders.createSimpleText(),
                 HELLO_WORLD);
