+++ tomcat/src/main/java/nextstep/jwp/controller/page/IndexController.java
@@ -7,6 +7,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import nextstep.jwp.controller.AbstractController;
 import nextstep.jwp.controller.Controller;
 import org.apache.coyote.http11.common.HttpHeaders;
 import org.apache.coyote.http11.common.HttpStatus;
@@ -14,17 +15,17 @@ import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.ResponseStatusLine;
 
-public class IndexGetPageController implements Controller {
+public class IndexController extends AbstractController {
 
-    private IndexGetPageController() {
+    private IndexController() {
     }
 
     public static Controller create() {
-        return new IndexGetPageController();
+        return new IndexController();
     }
 
     @Override
-    public HttpResponse process(final HttpRequest request) throws IOException {
+    protected HttpResponse doGet(final HttpRequest request) throws IOException {
         final String uri = request.getUri();
         final URL url = HttpResponse.class.getClassLoader()
                 .getResource(STATIC + uri);
